// src/pinecone-generated-ts/model/createRequestIndexConfig.ts
/**
 * Pinecone API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ApproximatedConfig } from './approximatedConfig';

export class CreateRequestIndexConfig {
    'kBits'?: number = 512;
    'hybrid'?: boolean = false;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "kBits",
            "baseName": "k_bits",
            "type": "number"
        },
        {
            "name": "hybrid",
            "baseName": "hybrid",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return CreateRequestIndexConfig.attributeTypeMap;
    }
}