// src/pinecone-generated-ts/model/deleteRequest.ts
/**
 * Pinecone API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* The request for the `Delete` operation.
*/
export class DeleteRequest {
    /**
    * Vectors to delete.
    */
    'ids'?: Array<string>;
    /**
    * This indicates that all vectors in the index namespace should be deleted.
    */
    'deleteAll'?: boolean = false;
    /**
    * The namespace to delete vectors from, if applicable.
    */
    'namespace'?: string;
    /**
    * If specified, the metadata filter here will be used to select the vectors to delete. This is mutually exclusive with specifying ids to delete in the ids param or using delete_all=True. See https://www.pinecone.io/docs/metadata-filtering/.
    */
    'filter'?: object;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "ids",
            "baseName": "ids",
            "type": "Array<string>"
        },
        {
            "name": "deleteAll",
            "baseName": "deleteAll",
            "type": "boolean"
        },
        {
            "name": "namespace",
            "baseName": "namespace",
            "type": "string"
        },
        {
            "name": "filter",
            "baseName": "filter",
            "type": "object"
        }    ];

    static getAttributeTypeMap() {
        return DeleteRequest.attributeTypeMap;
    }
}