// src/pinecone-generated-ts/model/fetchResponse.ts
/**
 * Pinecone API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Vector } from './vector';

/**
* The response for the `Fetch` operation.
*/
export class FetchResponse {
    'vectors'?: { [key: string]: Vector; };
    /**
    * The namespace of the vectors.
    */
    'namespace'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "vectors",
            "baseName": "vectors",
            "type": "{ [key: string]: Vector; }"
        },
        {
            "name": "namespace",
            "baseName": "namespace",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return FetchResponse.attributeTypeMap;
    }
}