// src/pinecone-generated-ts/model/hnswConfig.ts
/**
 * Pinecone API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class HnswConfig {
    'efConstruction'?: number = 500;
    'ef'?: number = 250;
    'm'?: number = 12;
    'maxElements'?: number = 50000000;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "efConstruction",
            "baseName": "ef_construction",
            "type": "number"
        },
        {
            "name": "ef",
            "baseName": "ef",
            "type": "number"
        },
        {
            "name": "m",
            "baseName": "M",
            "type": "number"
        },
        {
            "name": "maxElements",
            "baseName": "max_elements",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return HnswConfig.attributeTypeMap;
    }
}