// src/pinecone-generated-ts/model/indexMeta.ts
/**
 * Pinecone API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { IndexMetaDatabase } from './indexMetaDatabase';

export class IndexMeta {
    'database'?: IndexMetaDatabase;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "database",
            "baseName": "database",
            "type": "IndexMetaDatabase"
        }    ];

    static getAttributeTypeMap() {
        return IndexMeta.attributeTypeMap;
    }
}