// src/pinecone-generated-ts/model/indexMetaDatabase.ts
/**
 * Pinecone API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { IndexMetaDatabaseIndexConfig } from './indexMetaDatabaseIndexConfig';
import { IndexMetaDatabaseStatus } from './indexMetaDatabaseStatus';

export class IndexMetaDatabase {
    'name'?: string;
    'dimensions'?: string;
    'indexType'?: string;
    'metric'?: string;
    'pods'?: number;
    'replicas'?: number;
    'shards'?: number;
    'podType'?: string;
    'indexConfig'?: IndexMetaDatabaseIndexConfig;
    'status'?: IndexMetaDatabaseStatus;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "dimensions",
            "baseName": "dimensions",
            "type": "string"
        },
        {
            "name": "indexType",
            "baseName": "index_type",
            "type": "string"
        },
        {
            "name": "metric",
            "baseName": "metric",
            "type": "string"
        },
        {
            "name": "pods",
            "baseName": "pods",
            "type": "number"
        },
        {
            "name": "replicas",
            "baseName": "replicas",
            "type": "number"
        },
        {
            "name": "shards",
            "baseName": "shards",
            "type": "number"
        },
        {
            "name": "podType",
            "baseName": "pod_type",
            "type": "string"
        },
        {
            "name": "indexConfig",
            "baseName": "index_config",
            "type": "IndexMetaDatabaseIndexConfig"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "IndexMetaDatabaseStatus"
        }    ];

    static getAttributeTypeMap() {
        return IndexMetaDatabase.attributeTypeMap;
    }
}